+++ src/components/cart-icon/cart-icon.component.jsx
@@ -1,11 +1,15 @@
-import React, { useState } from "react";
+import React, { useContext } from "react";
+
+import { CartContext } from "../../context/cart.context";
 
 import { ReactComponent as ShoppingIcon } from "../../assets/shopping-bag.svg";
 
 import { CartIconContainer, ItemCount } from "./cart-icon.styles";
 
 const CartIcon = () => {
-  const [cartCount] = useState(0);
+  const { cart } = useContext(CartContext);
+
+  const cartCount = cart.reduce((acc, item) => acc + item.quantity, 0);
 
   return (
     <CartIconContainer>